Prevent QHttpRequest / QHttpResponse objects from being destroyed if the associated TCP connection is disconnected.
The documentation for QHttpServer states QHttpRequest objects are never deleted by QHttpServer and that
the QHttpResponse is only deleted when end() is called.

This is not the case however if the TCP connection is disconnected before the reply has been delivered -
since the QHttpRequest and QHttpResponse objects are owned by the QHttpConnection that creates them and
QHttpConnection is deleted if the TCP connection is disconnected.  This can happen if the server
'parks' the QHttpRequest/QHttpResponse objects whilst waiting for data from another source before
sending back the reply - in our case, user input.

This commit adds a ref count to QHttpConnection to track the count of active users (QHttpRequest or
QHttpResponse objects).

The QHttpConnection is released when there are no active users AND the TCP connection is disconnected.

With this change, the onus is on the user of QHttpServer to handle all requests, otherwise they will
be leaked.  Previously they would have been 'leaked' only until the connection was closed.