Revert "Prevent QHttpRequest / QHttpResponse objects from being destroyed if the associated TCP connection is disconnected."
This change caused two additional problems:

  * Any request/response objects not explicitly deleted are never freed which is a hassle for downstream
    consumers to have to deal with.

  * In addition to keeping the QHttpConnection alive, consumers also have to make sure they release
    any requests/responses before the QHttpServer is destroyed - since QHttpServer releases all connections
    when destroyed.

Instead, a simpler approach is to avoid holding on to dangling references downstream (eg. by using a QWeakPointer
or not holding on to a reference to QHttpRequest/QHttpResponse which could outlive the connection).

This commit instead keeps the old behavior but fixes the QHttpServer documentation about memory management.

This reverts commit caec4d345dacfd8bb69093759471422be75c06f1.